modify details

+++ server/src/main/java/org/apache/iotdb/db/qp/physical/sys/CreateAlignedTimeSeriesPlan.java
@@ -228,7 +228,7 @@ public class CreateAlignedTimeSeriesPlan extends PhysicalPlan {
     }
     compressors = new ArrayList<>();
     for (int i = 0; i < size; i++) {
-      compressors.add(CompressionType.deserialize((byte) buffer.get()));
+      compressors.add(CompressionType.deserialize(buffer.get()));
     }
 
     // alias
